==> 04-chap5.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Focus Group Discussion Text Analysis

```{r include=FALSE}
library(tidyverse)
library(tidytext)
library(wordcloud)
library(RColorBrewer)
library(gt)
library(patchwork)
library(scales)
library(quanteda)
library(corrplot)
```

```{r include=FALSE}
## Dataset
#restoration <- read_csv(here::here("data/GUI_NVivo_text_data.csv"))
values <- read_csv(here::here("data/2_GUI_VALUES.csv"))
benefits <- read_csv(here::here("data/2_GUI_BENEFITS.csv"))
```

```{r include=FALSE}
library("stopwords")
data("stop_words")
to_remove <- c(
  stopwords(language = "en", source = "smart"),
  "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um"
)
```

## Core Members VS Volunteers: Benefits 
### Restoration

```{r include=FALSE}
benefits %>% names()

benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("H : Restoration at GUI") %>%
  rename(Sentences = "H : Restoration at GUI")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("H : Restoration at GUI") %>%
  rename(Sentences = "H : Restoration at GUI")
```

```{r include=FALSE}
tidy_core_restoration <- benefits_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_restoration <- benefits_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_restoration_revise <- tidy_core_restoration %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_restoration_revise <- tidy_vol_restoration %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")


tidy_core_restoration %>%
  filter(!word
  %in% to_remove) %>%
  count(word, sort = TRUE) # %>% View()

tidy_vol_restoration %>%
  filter(!word
  %in% to_remove) %>%
  count(word, sort = TRUE) # %>% View()

tidy_core_restoration_revise$member <- "Core Members"
tidy_vol_restoration_revise$member <- "Volunteers"

tidy_core_restoration$member <- "Core Members"
tidy_vol_restoration$member <- "Volunteers"

combined_tidy_revise <- rbind(tidy_core_restoration_revise, tidy_vol_restoration_revise)
combined_tidy <- rbind(tidy_core_restoration, tidy_vol_restoration)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_restoration_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_restoration_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```



#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 5) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

- __Cosine Similarity__ is a metric used to determine how similar the documents are irrespective of their size.

- Cosine Similarity ranges from 0 to 1: 1 being the perfect match and 0 being totally different.

```{r include=FALSE}
#link: https://stackoverflow.com/questions/55629899/quanteda-calculate-text-similarity-by-row-between-two-dfms

tests <- benefits %>%
  select("H : Restoration at GUI") %>%
  rename(Sentences = "H : Restoration at GUI") %>%
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("H : Restoration at GUI") %>%
  rename(Sentences = "H : Restoration at GUI")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("H : Restoration at GUI") %>%
  rename(Sentences = "H : Restoration at GUI")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.585 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```


### Nature Connection

```{r include=FALSE}
benefits %>% names()

nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Nature Connection") %>%
  rename(Sentences = "G : Nature Connection")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Nature Connection") %>%
  rename(Sentences = "G : Nature Connection")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")


tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 4) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Nature Connection: Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("G : Nature Connection") %>%
  rename(Sentences = "G : Nature Connection") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Nature Connection") %>%
  rename(Sentences = "G : Nature Connection")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Nature Connection") %>%
  rename(Sentences = "G : Nature Connection")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.281 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE, fig.height=8, fig.width=8}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```



### Sense of Community

```{r include=FALSE}
nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("I : Sense of Community") %>%
  rename(Sentences = "I : Sense of Community")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("I : Sense of Community") %>%
  rename(Sentences = "I : Sense of Community")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")


tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 4) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Sense of Community: Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```



#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("I : Sense of Community") %>%
  rename(Sentences = "I : Sense of Community") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("I : Sense of Community") %>%
  rename(Sentences = "I : Sense of Community")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("I : Sense of Community") %>%
  rename(Sentences = "I : Sense of Community")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.696 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```


### Empowerment

```{r include=FALSE}
benefits %>% names()

nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Empowerment") %>%
  rename(Sentences = "A : Empowerment")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Empowerment") %>%
  rename(Sentences = "A : Empowerment")

tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 2) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Empowerment: Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("A : Empowerment") %>%
  rename(Sentences = "A : Empowerment") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Empowerment") %>%
  rename(Sentences = "A : Empowerment")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Empowerment") %>%
  rename(Sentences = "A : Empowerment")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.271 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```


### Environmental Awareness

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Environmental Awareness") %>%
  rename(Sentences = "B : Environmental Awareness")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Environmental Awareness") %>%
  rename(Sentences = "B : Environmental Awareness")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 1) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("B : Environmental Awareness") %>%
  rename(Sentences = "B : Environmental Awareness") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Environmental Awareness") %>%
  rename(Sentences = "B : Environmental Awareness")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Environmental Awareness") %>%
  rename(Sentences = "B : Environmental Awareness")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.151 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```




### Hands-on Experiences

```{r include=FALSE}
benefits %>% names()

nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Hands-on Experiences") %>%
  rename(Sentences = "C : Hands-on Experiences")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Hands-on Experiences") %>%
  rename(Sentences = "C : Hands-on Experiences")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 1) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```



#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("C : Hands-on Experiences") %>%
  rename(Sentences = "C : Hands-on Experiences") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Hands-on Experiences") %>%
  rename(Sentences = "C : Hands-on Experiences")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Hands-on Experiences") %>%
  rename(Sentences = "C : Hands-on Experiences")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.13 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```



### Mindsets (related to 5G)

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Mindsets (related to 5G)") %>%
  rename(Sentences = "D : Mindsets (related to 5G)")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Mindsets (related to 5G)") %>%
  rename(Sentences = "D : Mindsets (related to 5G)")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 2) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 10, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```





#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("D : Mindsets (related to 5G)") %>%
  rename(Sentences = "D : Mindsets (related to 5G)") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Mindsets (related to 5G)") %>%
  rename(Sentences = "D : Mindsets (related to 5G)")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
select("D : Mindsets (related to 5G)") %>%
  rename(Sentences = "D : Mindsets (related to 5G)")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.523 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```



### Self-awareness
No codes for self-awareness among the core members

### Self-Efficacy or Confidence
```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Self-Efficacy or Confidence") %>%
  rename(Sentences = "F : Self-Efficacy or Confidence")

nature_connection_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Self-Efficacy or Confidence") %>%
  rename(Sentences = "F : Self-Efficacy or Confidence")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 2) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 10, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```



#### Cosine Similarity

```{r include=FALSE}
tests <- benefits %>%
  select("F : Self-Efficacy or Confidence") %>%
  rename(Sentences = "F : Self-Efficacy or Confidence") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- benefits %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Self-Efficacy or Confidence") %>%
  rename(Sentences = "F : Self-Efficacy or Confidence")

benefits_vols <- benefits %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Self-Efficacy or Confidence") %>%
  rename(Sentences = "F : Self-Efficacy or Confidence")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.395 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```


## Core Members VS Volunteers: Values 
### Characteristics of Nature in GUI

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Characteristics of Nature in GUI") %>%
  rename(Sentences = "B : Characteristics of Nature in GUI")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Characteristics of Nature in GUI") %>%
  rename(Sentences = "B : Characteristics of Nature in GUI")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 2) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("B : Characteristics of Nature in GUI") %>%
  rename(Sentences = "B : Characteristics of Nature in GUI") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Characteristics of Nature in GUI") %>%
  rename(Sentences = "B : Characteristics of Nature in GUI")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("B : Characteristics of Nature in GUI") %>%
  rename(Sentences = "B : Characteristics of Nature in GUI")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.53 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```

### Sense of Autonomy and Ownership

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Sense of Autonomy and Ownership") %>%
  rename(Sentences = "F : Sense of Autonomy and Ownership")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Sense of Autonomy and Ownership") %>%
  rename(Sentences = "F : Sense of Autonomy and Ownership")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 2) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 10, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("F : Sense of Autonomy and Ownership") %>%
  rename(Sentences = "F : Sense of Autonomy and Ownership") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Sense of Autonomy and Ownership") %>%
  rename(Sentences = "F : Sense of Autonomy and Ownership")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("F : Sense of Autonomy and Ownership") %>%
  rename(Sentences = "F : Sense of Autonomy and Ownership")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.548 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200),
         cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```



### Risk-taking but Safe Space

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("E : Risk-taking but Safe Space") %>%
  rename(Sentences = "E : Risk-taking but Safe Space")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("E : Risk-taking but Safe Space") %>%
  rename(Sentences = "E : Risk-taking but Safe Space")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 4) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("E : Risk-taking but Safe Space") %>%
  rename(Sentences = "E : Risk-taking but Safe Space") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("E : Risk-taking but Safe Space") %>%
  rename(Sentences = "E : Risk-taking but Safe Space")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("E : Risk-taking but Safe Space") %>%
  rename(Sentences = "E : Risk-taking but Safe Space")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.589 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200),
         cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```



### Connecting people

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Connecting people") %>%
  rename(Sentences = "C : Connecting people")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Connecting people") %>%
  rename(Sentences = "C : Connecting people")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 3) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("C : Connecting people") %>%
  rename(Sentences = "C : Connecting people") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Connecting people") %>%
  rename(Sentences = "C : Connecting people")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("C : Connecting people") %>%
  rename(Sentences = "C : Connecting people")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.502 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200),
         cl.pos = "n",
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```





### Activism

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Activism") %>%
  rename(Sentences = "A : Activism")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Activism") %>%
  rename(Sentences = "A : Activism")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 1) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 5, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```



#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("A : Activism") %>%
  rename(Sentences = "A : Activism") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Activism") %>%
  rename(Sentences = "A : Activism")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("A : Activism") %>%
  rename(Sentences = "A : Activism")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.281 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n", 
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```



### Non hierarchical Structure

```{r include=FALSE}
benefits %>% names()
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Non hierarchical Structure") %>%
  rename(Sentences = "D : Non hierarchical Structure")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Non hierarchical Structure") %>%
  rename(Sentences = "D : Non hierarchical Structure")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 2) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 10, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```


#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("D : Non hierarchical Structure") %>%
  rename(Sentences = "D : Non hierarchical Structure") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Non hierarchical Structure") %>%
  rename(Sentences = "D : Non hierarchical Structure")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("D : Non hierarchical Structure") %>%
  rename(Sentences = "D : Non hierarchical Structure")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.105 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n", 
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```





### Sense of Place

```{r include=FALSE}
values %>% names()

nature_connection_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Sense of Place") %>%
  rename(Sentences = "G : Sense of Place")

nature_connection_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Sense of Place") %>%
  rename(Sentences = "G : Sense of Place")
```

```{r include=FALSE}
tidy_core_n_connection <- nature_connection_core %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

tidy_vol_n_connection <- nature_connection_vols %>%
  unnest_tokens(word, Sentences) # unesting sentensec into words

data(stop_words)

tidy_core_n_connection_revise <- tidy_core_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_vol_n_connection_revise <- tidy_vol_n_connection %>%
  filter(!word %in% to_remove)
# anti_join(stop_words, by="word")

tidy_core_n_connection_revise$member <- "Core Members"
tidy_vol_n_connection_revise$member <- "Volunteers"

combined_tidy_connect_revise <- rbind(tidy_core_n_connection_revise, tidy_vol_n_connection_revise)
```

#### Most Frequently Used Words
```{r echo=FALSE}
a1 <- tidy_core_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkgreen") +
  labs(x = "Words", y = "", title = "By Core Members") +
  coord_flip() +
  theme_minimal()

a2 <- tidy_vol_n_connection_revise %>%
  group_by(word) %>%
  summarise(count = n()) %>%
  arrange(-count) %>%
  head(10) %>%
  ggplot() +
  geom_col(aes(x = reorder(word, count), y = count), fill = "darkorange") +
  labs(x = "Words", y = "", title = "By Volunteers") +
  coord_flip() +
  theme_minimal()

a1 | a2
```

#### Frequency–inverse Document Frequency
This is to find words that are common within the messages of one author but uncommon in the rest of the messages.

```{r echo=FALSE}
combined_tidy_connect_revise %>%
  filter(!word %in% to_remove) %>%
  count(member, word, sort = TRUE) %>%
  bind_tf_idf(term = word, document = member, n = n) %>%
  filter(n > 5) %>%
  group_by(member) %>%
  # filter(tf_idf>=0.0002149294) %>%
  top_n(n = 10, tf_idf) %>%
  ggplot(aes(x = reorder_within(word, n, member), y = n, fill = member)) +
  geom_col(show.legend = FALSE) +
  ylab("") +
  xlab("") +
  coord_flip() +
  facet_wrap(~member, ncol = 2, scales = "free_y") +
  scale_x_reordered() +
  ggtitle("Unique words used by Core members VS Volunteers") +
  theme(
    axis.ticks.y = element_blank(),
    # text=element_text(family = font),
    plot.title = element_text(
      face = "bold",
      # colour = colortitles,
      size = 13, hjust = 0.5
    )
  )
```





#### Cosine Similarity

```{r include=FALSE}
tests <- values %>%
  select("G : Sense of Place") %>%
  rename(Sentences = "G : Sense of Place") %>% 
  rename(text = "Sentences") %>%
  mutate(id = row_number() #,
         # source = case_when(
         #   id == 1 ~ "Core Members",
         #   id == 2 ~ "Core Members",
         #   TRUE ~ "Volunteers")
         ) %>%
  select(id, text, #source
         everything())

## Volunteers vs Core
benefits_core <- values %>%
  filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Sense of Place") %>%
  rename(Sentences = "G : Sense of Place")

benefits_vols <- values %>%
  filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>%
  select("G : Sense of Place") %>%
  rename(Sentences = "G : Sense of Place")

benefits_core <- paste(unlist(benefits_core), collapse =" ")
benefits_vols <- paste(unlist(benefits_vols), collapse =" ")

test_3 <- data.frame(id=c(1, 2))

test_3$text <- c(benefits_core, benefits_vols)

# test_2<- data.frame(id=c(1, 2),
#            text=c(benefits_core, benefits_vols))

test_3 <- data.frame(test_3)

test_4 <- test_3 %>%
  filter(id == 1)
test_4$id <- 0

vols <- tests %>% filter(id !=1 & id !=2)

vol_core <- data.frame(rbind(test_4, vols))
```

- Overall __Cosine Similarity__ between Core members and Volunteers

```{r echo=FALSE}
corp_2 <- corpus(test_3)
#docnames(corp_2) <- paste(df$source, sep = "_")
dfm(corp_2, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")) %>% #View()
  textstat_simil(method = "cosine") #0.308 overall
```

- __Cosine Similarity__ between Core members (text 1) and 5 focus group sessions with volunteers (text 2 to 6)

```{r echo=FALSE}
corp <- corpus(vol_core)
#docnames(corp) <- paste(df$source, sep = "_")
cor_M<- dfm(corp, remove_punct = TRUE , remove = c(stopwords("english"), "yeah", "uh", "it’s", "you’re", "there’s", "i’m", "don’t", "i’ve", "you’ve", "that’s", "things", "cos", "er", "erm", "everyone’s", "lo", "lor", "ah", "we’re", "um", "also")
    ) %>%
  textstat_simil(method = "cosine")
#0.194 to 0.426

cor_M <- as.matrix(cor_M)

col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
corrplot(cor_M, method = "color", col = col(200), cl.pos = "n", 
         type = "upper", number.cex = .7,
         addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 45, # Text label color and rotation
         # Combine with significance
         #p.mat = confident_95$p, sig.level = 0.05, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)

```









## Summary

### Comparison between Core Members and Volunteers

#### Benefits

- Relatively high `Cosine Similarity` are found among `Restoration`, `Nature Connection`, `Sense of Community`, and `Mindsets (related to 5G)` between GUI Core Members and volunteers.

- It is interesting to note that core members and volunteers seem to use different words to describe `Empowerment`, `Environmental Awareness`, `Hands-on Experiences`, and `Self-Efficacy or Confidence`.


```{r echo=FALSE}
tibble(
  "GUI Benefits" = c(
    "Restoration", "Nature Connection", "Sense of Community", "Empowerment", "Environmental Awareness", "Hands-on Experiences", "Mindsets (related to 5G)", "Self-Confidence", "Self-awareness"
  ),
  "Cosine Similarity" = c(0.586, 0.629, 0.696, 0.271, 0.151, "0.130", 0.523, 0.395, "NA*")
) %>%
  gt() %>%
  tab_header(
    title = "GUI Benefits: Comparison between Core Members and Volunteers"
  ) %>%
  tab_source_note(
    source_note = "*Self-awareness was not found among core members."
  ) %>% 
  cols_align(align = "left") 

```

#### Values

-Relatively high `Cosine Similarity` are found among `Characteristics of Nature in GUI`, `Sense of Autonomy and Ownership`, `Risk-taking but Safe Space`, and `Connecting people`.

- For `Characteristics of Nature in GUI`, `Activism`, `Non hierarchical Structure`, and `Sense of Place`, core members and volunteers seem to use different words to describe them.


```{r echo=FALSE}
tibble(
  "GUI Values" = c(
    "Characteristics of Nature in GUI", "Sense of Autonomy and Ownership",
    "Risk-taking but Safe Space", "Connecting people", "Activism", "Non hierarchical Structure", "Sense of Place"
  ),
  "Cosine Similarity" = c(0.530, 0.548, 0.589, 0.502, 0.281, 0.105, 0.308)
) %>%
  gt() %>%
  tab_header(
    title = "GUI Values: Comparison between Core Members and Volunteers"
  ) %>% 
  cols_align(align = "left") 
# %>%
#   tab_source_note(
#     source_note = "* denotes non significant correlation"
#   )
```

<!-- ### Comparison across Volunteer Groups and Core Members -->

<!-- #### Benefits -->
<!-- - As observed in the comparison between core members and volunteers, significant correlation on the word usages was identified in the __`Restoration`(group 2 & 3, 2 & 4, and 2 & 5), `Nature Connection` (group 1 & 3, 1 & 4, and 3 & 4), `Sense of Community` (all groups), and `Mindsets (related to 5G)` (group 1 & 3, 2 & 4, and 2 & 5)__ across the different focus group sessions -->

<!-- - Even compared to the core group, most of the groups in these four benefits show __significant correlations__: `Restoration`(group 1, 2, 3 and 5), `Nature Connection` (all groups), `Sense of Community` (all groups), and `Mindsets (related to 5G)` (group 1, 2, 4 and 5) -->

<!-- - The word usage was especially __consistent (strong significant correlation)__ in the __`Sense of Community`__ across __all the five sessions__ -->

<!-- __Restoration__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("H : Restoration at GUI") %>% -->
<!--   rename(Sentences = "H : Restoration at GUI") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("H : Restoration at GUI") %>% -->
<!--   rename(Sentences = "H : Restoration at GUI") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->
<!-- # frequency_revise <- bind_rows( -->
<!-- #   mutate(tidy_core_restoration_revise, member = "Core Members"), -->
<!-- #   mutate(tidy_vol_restoration_revise, member = "Volunteers") -->
<!-- # ) %>% -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- #Use this version -->
<!-- library(corrplot) -->
<!-- #write_csv(tidy_bene_vols, ("data/tidy_bene_vols.csv")) -->

<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Nature Connection__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Nature Connection") %>% -->
<!--   rename(Sentences = "G : Nature Connection") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Nature Connection") %>% -->
<!--   rename(Sentences = "G : Nature Connection") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- library(corrplot) -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Sense of Community__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("I : Sense of Community") %>% -->
<!--   rename(Sentences = "I : Sense of Community") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--    select("I : Sense of Community") %>% -->
<!--   rename(Sentences = "I : Sense of Community") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Empowerment__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Empowerment") %>% -->
<!--   rename(Sentences = "A : Empowerment") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Empowerment") %>% -->
<!--   rename(Sentences = "A : Empowerment") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Environmental Awareness__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Environmental Awareness") %>% -->
<!--   rename(Sentences = "B : Environmental Awareness") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Environmental Awareness") %>% -->
<!--   rename(Sentences = "B : Environmental Awareness") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Hands-on Experiences__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("C : Hands-on Experiences") %>% -->
<!--   rename(Sentences = "C : Hands-on Experiences") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("C : Hands-on Experiences") %>% -->
<!--   rename(Sentences = "C : Hands-on Experiences") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->
<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Mindsets (related to 5G)__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Mindsets (related to 5G)") %>% -->
<!--   rename(Sentences = "D : Mindsets (related to 5G)") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Mindsets (related to 5G)") %>% -->
<!--   rename(Sentences = "D : Mindsets (related to 5G)") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->
<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Self-Efficacy or Confidence__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Self-Efficacy or Confidence") %>% -->
<!--   rename(Sentences = "F : Self-Efficacy or Confidence") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Self-Efficacy or Confidence") %>% -->
<!--   rename(Sentences = "F : Self-Efficacy or Confidence") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- library(corrplot) -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->


<!-- #### Values -->

<!-- - Constrast to the comparison of core members and volunteers, __only a few values__ show significant correlations across the five groups of volunteers -->

<!-- - Two values show __more consistency (significant correlations)__, including `Sense of Autonomy and Ownership` (group 2 & 5, 3 & 5, and 4 & 5), `Risk-taking but Safe Space` (group 2 & 3, 3 & 4, 3 & 5, and, 4 & 5), and `Connecting people` (group 1 & 5, 2 & 3, 2 & 4, 2 & 5, and 4 & 5)  -->

<!-- - Significant correlations were observed only among the selected groups in `Activism`  (group 2 & 5) -->

<!-- - Compared to the core group, most of the groups in these three values show __significant correlations__: -->
<!-- `Sense of Autonomy and Ownership` (group 2, 4 and 5), `Risk-taking but Safe Space` (group 3, 4, and 5), and `Connecting people` (group 2 and 5)  -->


<!-- __Characteristics of Nature in GUI__ -->

<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Characteristics of Nature in GUI") %>% -->
<!--   rename(Sentences = "B : Characteristics of Nature in GUI")  -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Characteristics of Nature in GUI") %>% -->
<!--   rename(Sentences = "B : Characteristics of Nature in GUI") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Sense of Autonomy and Ownership__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Sense of Autonomy and Ownership") %>% -->
<!--   rename(Sentences = "F : Sense of Autonomy and Ownership") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Sense of Autonomy and Ownership") %>% -->
<!--   rename(Sentences = "F : Sense of Autonomy and Ownership") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->
<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Risk-taking but Safe Space__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("E : Risk-taking but Safe Space") %>% -->
<!--   rename(Sentences = "E : Risk-taking but Safe Space") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("E : Risk-taking but Safe Space") %>% -->
<!--   rename(Sentences = "E : Risk-taking but Safe Space") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Connecting people__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--    select("C : Connecting people") %>% -->
<!--   rename(Sentences = "C : Connecting people") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("C : Connecting people") %>% -->
<!--   rename(Sentences = "C : Connecting people") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Activism__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Activism") %>% -->
<!--   rename(Sentences = "A : Activism") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Activism") %>% -->
<!--   rename(Sentences = "A : Activism") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Non hierarchical Structure__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Non hierarchical Structure") %>% -->
<!--   rename(Sentences = "D : Non hierarchical Structure") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Non hierarchical Structure") %>% -->
<!--   rename(Sentences = "D : Non hierarchical Structure") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Sense of Place__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Sense of Place") %>% -->
<!--   rename(Sentences = "G : Sense of Place") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Sense of Place") %>% -->
<!--   rename(Sentences = "G : Sense of Place") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->
