==> 05-chap6.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary of Findings

```{r include=FALSE}
library(tidyverse)
library(psych)
library(lsr)
library(gt)
library(tidyr)
library(tidytext)
library(corrplot)
library(patchwork)
```

## Research Question 1

__*How nature place-making unfolds when conducted by NGOs*__

*Qualitative analysis of Focus Group Discussions will be added.


## Research Question 2

__*To what extent the facilitators of nature place-making conducted by NGOs influence the individual’s self-identity and perception of the place*__


### Cosine Similarity: Comparison between Core Members and Volunteers

- __Cosine similarity__ is a metric used to determine __how similar the documents__ are irrespective of their size.

- Mathematically, it measures the cosine of the angle between two vectors projected in a multi-dimensional space. In this context, the two vectors are arrays containing the word counts of two documents.

- When plotted on a multi-dimensional space, where each dimension corresponds to a word in the document, the cosine similarity captures the __orientation (the angle)__ of the documents and not the magnitude. 

- This method is based on an assumption which is __the document (or sentence) are similar if having common words__.


#### Benefits

- Relatively __high `Cosine Similarity`__ are found among __`Restoration`, `Nature Connection`, `Sense of Community`, and `Mindsets (related to 5G)`__ between GUI Core Members and volunteers.

- It is interesting to note that core members and volunteers seem to use different words to describe __`Empowerment`, `Environmental Awareness`, `Hands-on Experiences`, and `Self-Efficacy or Confidence`__.

```{r echo=FALSE}
tibble(
  "GUI Benefits" = c(
    "Restoration", "Nature Connection", "Sense of Community", "Empowerment", "Environmental Awareness", "Hands-on Experiences", "Mindsets (related to 5G)", "Self-awareness", "Self-Efficacy or Confidence"
  ),
  "Cosine Similarity" = c(0.586, 0.629, 0.696, 0.271, 0.151, 0.13, 0.523, "NA*", 0.395)
) %>%
  gt() %>%
  tab_header(
    title = "GUI Benefits: Comparison between Core Members and Volunteers"
  ) %>%
  tab_source_note(
    source_note = "*Self-awareness was not found among core members."
  )
```


#### Values

- Relatively __high `Cosine Similarity`__ are found among __`Characteristics of Nature in GUI`, `Sense of Autonomy and Ownership`, `Risk-taking but Safe Space`, and `Connecting people`__.

- For `__Characteristics of Nature in GUI`, `Activism`, `Non hierarchical Structure`, and `Sense of Place`__, core members and volunteers seem to __use different words__ to describe them.

```{r echo=FALSE}
tibble(
  "GUI Values" = c(
    "Characteristics of Nature in GUI", "Sense of Autonomy and Ownership",
    "Risk-taking but Safe Space", "Connecting people", "Activism", "Non hierarchical Structure", "Sense of Place*"
  ),
  "Cosine Similarity" = c(0.530, 0.548, 0.589, 0.502, 0.281, 0.105, 0.308)
) %>%
  gt() %>%
  tab_header(
    title = "GUI Values: Comparison between Core Members and Volunteers"
  ) 
# %>%
#   tab_source_note(
#     source_note = "* denotes non significant correlation"
#   )
```


<!-- __Restoration__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("H : Restoration at GUI") %>% -->
<!--   rename(Sentences = "H : Restoration at GUI") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("H : Restoration at GUI") %>% -->
<!--   rename(Sentences = "H : Restoration at GUI") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->
<!-- # frequency_revise <- bind_rows( -->
<!-- #   mutate(tidy_core_restoration_revise, member = "Core Members"), -->
<!-- #   mutate(tidy_vol_restoration_revise, member = "Volunteers") -->
<!-- # ) %>% -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- #Use this version -->
<!-- #write_csv(tidy_bene_vols, ("data/tidy_bene_vols.csv")) -->

<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Nature Connection__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Nature Connection") %>% -->
<!--   rename(Sentences = "G : Nature Connection") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Nature Connection") %>% -->
<!--   rename(Sentences = "G : Nature Connection") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- library(corrplot) -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Sense of Community__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("I : Sense of Community") %>% -->
<!--   rename(Sentences = "I : Sense of Community") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--    select("I : Sense of Community") %>% -->
<!--   rename(Sentences = "I : Sense of Community") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Empowerment__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Empowerment") %>% -->
<!--   rename(Sentences = "A : Empowerment") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Empowerment") %>% -->
<!--   rename(Sentences = "A : Empowerment") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Environmental Awareness__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Environmental Awareness") %>% -->
<!--   rename(Sentences = "B : Environmental Awareness") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Environmental Awareness") %>% -->
<!--   rename(Sentences = "B : Environmental Awareness") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Hands-on Experiences__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("C : Hands-on Experiences") %>% -->
<!--   rename(Sentences = "C : Hands-on Experiences") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("C : Hands-on Experiences") %>% -->
<!--   rename(Sentences = "C : Hands-on Experiences") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->
<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Mindsets (related to 5G)__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Mindsets (related to 5G)") %>% -->
<!--   rename(Sentences = "D : Mindsets (related to 5G)") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Mindsets (related to 5G)") %>% -->
<!--   rename(Sentences = "D : Mindsets (related to 5G)") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->
<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Self-Efficacy or Confidence__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- benefits %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Self-Efficacy or Confidence") %>% -->
<!--   rename(Sentences = "F : Self-Efficacy or Confidence") -->

<!-- bene_vols <- benefits %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Self-Efficacy or Confidence") %>% -->
<!--   rename(Sentences = "F : Self-Efficacy or Confidence") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->


<!-- #### Values -->

<!-- - Constrast to the comparison of core members and volunteers, __only a few values__ show significant correlations across the five groups of volunteers -->

<!-- - Two values show __more consistency (significant correlations)__, including `Sense of Autonomy and Ownership` (group 2 & 5, 3 & 5, and 4 & 5), `Risk-taking but Safe Space` (group 2 & 3, 3 & 4, 3 & 5, and, 4 & 5), and `Connecting people` (group 1 & 5, 2 & 3, 2 & 4, 2 & 5, and 4 & 5)  -->

<!-- - Significant correlations were observed only among the selected groups in `Activism`  (group 2 & 5) -->

<!-- - Compared to the core group, most of the groups in these three values show __significant correlations__: -->
<!-- `Sense of Autonomy and Ownership` (group 2, 4 and 5), `Risk-taking but Safe Space` (group 3, 4, and 5), and `Connecting people` (group 2 and 5)  -->


<!-- __Characteristics of Nature in GUI__ -->

<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Characteristics of Nature in GUI") %>% -->
<!--   rename(Sentences = "B : Characteristics of Nature in GUI")  -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("B : Characteristics of Nature in GUI") %>% -->
<!--   rename(Sentences = "B : Characteristics of Nature in GUI") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Sense of Autonomy and Ownership__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Sense of Autonomy and Ownership") %>% -->
<!--   rename(Sentences = "F : Sense of Autonomy and Ownership") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("F : Sense of Autonomy and Ownership") %>% -->
<!--   rename(Sentences = "F : Sense of Autonomy and Ownership") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->
<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Risk-taking but Safe Space__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("E : Risk-taking but Safe Space") %>% -->
<!--   rename(Sentences = "E : Risk-taking but Safe Space") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("E : Risk-taking but Safe Space") %>% -->
<!--   rename(Sentences = "E : Risk-taking but Safe Space") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Connecting people__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--    select("C : Connecting people") %>% -->
<!--   rename(Sentences = "C : Connecting people") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("C : Connecting people") %>% -->
<!--   rename(Sentences = "C : Connecting people") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Activism__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Activism") %>% -->
<!--   rename(Sentences = "A : Activism") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("A : Activism") %>% -->
<!--   rename(Sentences = "A : Activism") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Non hierarchical Structure__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Non hierarchical Structure") %>% -->
<!--   rename(Sentences = "D : Non hierarchical Structure") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("D : Non hierarchical Structure") %>% -->
<!--   rename(Sentences = "D : Non hierarchical Structure") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->

<!-- __Sense of Place__ -->
<!-- ```{r include=FALSE} -->
<!-- bene_core <- values %>% -->
<!--   filter(Workshop == "1 : 1_GUI_FGD_1_Core_1_" | Workshop == "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Sense of Place") %>% -->
<!--   rename(Sentences = "G : Sense of Place") -->

<!-- bene_vols <- values %>% -->
<!--   filter(Workshop != "1 : 1_GUI_FGD_1_Core_1_" & Workshop != "2 : 2_GUI_FGD_2_Core_2") %>% -->
<!--   select("G : Sense of Place") %>% -->
<!--   rename(Sentences = "G : Sense of Place") %>%  -->
<!--   mutate(workshops =row_number()) %>%  -->
<!--   select(workshops, everything()) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- tidy_bene_vols <- bene_vols %>% -->
<!--   unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) -->

<!-- tidy_bene_core <- bene_core %>%  -->
<!-- unnest_tokens(word, Sentences) %>%  # unesting sentensec into words -->
<!--   filter(!word %in% to_remove) %>%  -->
<!--   mutate(workshops = "Core") %>%  -->
<!--   select(workshops, everything()) -->

<!-- tidy_bene_vols <- rbind(tidy_bene_vols, tidy_bene_core) -->
<!-- ``` -->

<!-- ```{r include=FALSE} -->
<!-- library(tidyr) -->
<!-- na_strings <- c("NA", "n/a") -->

<!-- tidy_bene_vols <- tidy_bene_vols %>%  -->
<!--   mutate(word = str_extract(word, "[a-z']+")) %>% -->
<!--   count(workshops, word) %>% -->
<!--   group_by(workshops) %>% -->
<!--   mutate(proportion = n / sum(n)) %>% -->
<!--   select(-n) %>% -->
<!--   spread(workshops, proportion) %>% -->
<!--   naniar::replace_with_na_all(condition = ~ . %in% na_strings) %>% -->
<!--   filter(word != "NA") -->

<!-- # frequency_revise %>% select(`Core Members`) %>% is.na(`Core Members`) -->

<!-- tidy_bene_vols[is.na(tidy_bene_vols)] <- 0 -->
<!-- # glimpse(frequency_2) -->

<!-- tidy_bene_vols %>% -->
<!--   select(-word) %>% -->
<!--   pairs() -->

<!-- tidy_bene_vols <- data.frame(tidy_bene_vols) -->
<!-- ``` -->

<!-- ```{r echo=FALSE} -->
<!-- confident_95 <- cor.mtest( tidy_bene_vols %>% select(-word), conf.level = .95) -->

<!-- cor_M <- tidy_bene_vols %>%  -->
<!--   select(-word) %>% -->
<!--   cor() -->

<!-- col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444")) -->
<!-- corrplot(cor_M, method = "color", col = col(200), -->
<!--          type = "upper", number.cex = .7, -->
<!--          addCoef.col = "black", # Add coefficient of correlation -->
<!--          tl.col = "black", tl.srt = 90, # Text label color and rotation -->
<!--          # Combine with significance -->
<!--          p.mat = confident_95$p, sig.level = 0.05, insig = "blank",  -->
<!--          # hide correlation coefficient on the principal diagonal -->
<!--          diag = FALSE) -->
<!-- ``` -->



## Research Question 3

__*How nature place-making activities impact the subjective sense of well-being of individuals participating in the activities*__

### Comparing GUI's fresh and long-term members

- Although the ANOVA test did not find any sigfnificant differences in the psychometric scales across the groups, the difference in the median of __Sense of Community__ between the __fresh_members__ and __long term members__ is hgih. Hence, we ran the Median Test.

- A __significant__ difference ( _p_ =0.014) was found in __Sense of Community__ between __Fresh Members (`1`)__ and __Long-term GUI Members (`2`)__: hence, the __long-term engagement with GUI__ seems to __enhance a Sense of Community__ of GUI members.

- __Committed frequent visitors (`3`)__ and __Everyday visitors (`4`)__ did not show a significant difference in __Sense of Community__ when it's compared with __fresh_members__. Perhaps, this is due to the small number of group size (7 and 2, respectively). 

- __Self-esteem (`Self_Est`), Intrinsic Motivation (`IM`), Perceived Competence (`PComp`), Perceived Choice (`PChoice`), Self Efficacy (`Self_Eff`), and Self Esteem (`Self_Est`)__ do not seem to be developed through over time and frequent interactions

```{r include=FALSE}
gui_cluster_data <- readRDS(here::here("data/only_gui_cluster_June4.rds"))
cluster_wise <- gui_cluster_data %>%
  select(age, frequency_coded, commitment_coded, number_programs, SoCoh, SOC, IMI, IM, PComp, PChoice, CNS, Self_Est, Self_Eff, clusters) %>%
  describeBy(., group = "clusters")

group1 <- round(cluster_wise$`1`, 2)
group1_rev <- group1 %>%
  select(median) %>%
  t() %>%
  data.frame()

group2 <- round(cluster_wise$`2`, 2)
group2_rev <- group2 %>%
  select(median) %>%
  t() %>%
  data.frame()

group3 <- round(cluster_wise$`3`, 2)
group3_rev <- group3 %>%
  select(median) %>%
  t() %>%
  data.frame()

group4 <- round(cluster_wise$`4`, 2)
group4_rev <- group4 %>%
  select(median) %>%
  t() %>%
  data.frame()

median_clusters <- rbind(group1_rev, group2_rev, group3_rev, group4_rev) %>% select(clusters., everything())

gui_cluster_data %>%
  select(clusters) %>%
  group_by(clusters) %>%
  summarise(count = n()) %>%
  gt()

# Mean Values
# group11 <- round(cluster_wise$`1`, 2)
# group11_rev <- group11 %>% select(mean) %>% t() %>% data.frame()
#
# group22 <- round(cluster_wise$`2`, 2)
# group22_rev <- group22 %>% select(mean) %>% t() %>% data.frame()
#
# group33 <- round(cluster_wise$`3`, 2)
# group33_rev <- group33 %>% select(mean) %>% t() %>% data.frame()
#
# group44 <- round(cluster_wise$`4`, 2)
# group44_rev <- group44 %>% select(mean) %>% t() %>% data.frame()
#
# mean_clusters<- rbind(group11_rev, group22_rev, group33_rev, group44_rev) %>% select(clusters., everything())
#
# mean_clusters %>% gt() %>% tab_header(
#     title = "Mean Values by Groups"
#   ) %>%
#   cols_align(align = "center")

gui_cluster_data %>%
  filter(clusters == 1) %>%
  group_by(gender) %>%
  summarize(count = n())

gui_cluster_data %>%
  filter(clusters == 4) %>%
  group_by(gender) %>%
  summarize(count = n())
```

```{r echo=FALSE}
median_clusters %>%
  mutate(
    male = c(17, 7, 1, 3),
    female = c(49, 21, 1, 4),
    members_n = c(67, 28, 2, 7)
  ) %>%
  select(clusters., male, female, members_n, everything()) %>%
  gt() %>%
  tab_header(
    title = "Table A: Median Values by Groups"
  ) %>%
  cols_align(align = "center")
```

```{r echo=FALSE}
c1 <- gui_cluster_data %>% 
  #filter(clusters != 3 & clusters != 4) %>% 
  select(CNS, clusters) %>%
  ggplot(aes(x = clusters, y = CNS, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")


c2 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>% 
  select(SOC, clusters) %>%
  ggplot(aes(x = clusters, y = SOC, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

c3 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(SoCoh, clusters) %>%
  ggplot(aes(x = clusters, y = SoCoh, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

c4 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(IMI, clusters) %>%
  ggplot(aes(x = clusters, y = IMI, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

e4 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(IM, clusters) %>%
  ggplot(aes(x = clusters, y = IM, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

e5 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(PComp, clusters) %>%
  ggplot(aes(x = clusters, y = PComp, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

e6 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(PChoice, clusters) %>%
  ggplot(aes(x = clusters, y = PChoice, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

c5 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(Self_Eff, clusters) %>%
  ggplot(aes(x = clusters, y = Self_Eff, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

c6 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(Self_Est, clusters) %>%
  ggplot(aes(x = clusters, y = Self_Est, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

c7 <- gui_cluster_data %>%
  #filter(clusters != 3 & clusters != 4) %>%
  select(age, clusters) %>%
  ggplot(aes(x = clusters, y = age, group = clusters, fill = clusters)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position = "none")

c1 + c2 + c3 + c4 + e4 + e5 + e6 + c5 + c6 + c7
```

```{r include=FALSE}
# t test and Effect Size (Gender) for IMI and SOC
gui_cluster_data <- readRDS(here::here("data/only_gui_cluster_June4.rds"))
old_new <- gui_cluster_data %>%
  filter(clusters == 1 | clusters == 2)
library(agricolae)
```

```{r include=FALSE}
Median.test(old_new$SOC, old_new$clusters) # Sense of Community is significant. P=0.014
Median.test(old_new$CNS, old_new$clusters) # CNS is not significant. p=0.374
Median.test(old_new$IMI, old_new$clusters) # IMI is not significant. p=0.491
Median.test(old_new$SoCoh, old_new$clusters) # Social Cohesion is not significant. p=0.333
Median.test(old_new$Self_Est, old_new$clusters) # Self-est is not significant. p=0.433
Median.test(old_new$Self_Eff, old_new$clusters) # Self-eff is not significant. p=0.494
Median.test(old_new$IM, old_new$clusters) # IM is not significant. p=0.64
Median.test(old_new$PComp, old_new$clusters) # PComp is not significant. p=0.947
Median.test(old_new$PChoice, old_new$clusters) # PChoice is not significant. p=0.739
```

### Short-Term Effects of attending GUI Program

- We conducted the __Welch Two Sample t-test__ on `PANAS` and __Wilcoxon Signed-rank test__ on `ROS`.

- __Statistically significant differences__ were observed in __both scales__. Cohen's d was estimated at __0.51 and 1.002__, which are __medium to large effect__ based on the Cohen’s guidelines (1992).

- This study identified that __those who engaged in GUI programs__ feel __more restored and more positive emmotions__ right __after attending a program__. 

- Going forward, __program-wise analysis__ needs to be conducted to see if __attending a particular program__ shows __different ROS and PANAS__ compared to those who attend a different program. For this, we need further data collections (minimum of 20 participants in each program).

```{r include=FALSE}
data_pre_post_scales <- readRDS(here::here("data/data_pre_post_scales.rds"))

# t test for SOC and Effect Size
t.test(data_pre_post_scales$PANAS ~ data_pre_post_scales$pre_post, mu = 0, alt = "two.sided", conf = 0.95, var.eq = F, paired = T)

library(effsize)
cohen.d(data_pre_post_scales$PANAS ~ data_pre_post_scales$pre_post, na.rm = TRUE, pooled = TRUE, paired = T)

wilcox.test(data_pre_post_scales$ROS ~ data_pre_post_scales$pre_post, mu = 0, alt = "two.sided", conf.int = T, conf.level = 0.95, paired = T, exact = T, correct = T)

library(rstatix)
data_pre_post_scales %>%
  wilcox_effsize(ROS ~ pre_post, paired = T)

data_pre_post_scales %>%
  select(ROS, PANAS) %>%
  describeBy(group = data_pre_post_scales$pre_post)

pp_1 <- tibble(vars = c(1, 2), scales = c("ROS", "PANAS"), Cronbachs_alpha = c(0.88, 0.90))

pp_2 <- pp_1 %>% # Cronbacks Alpha
  mutate(
    shapiro_test = c(0.018, 0.052), # Normality shapiro test
    pre_means = c(4.94, 5.20), # pre_means
    post_means = c(5.78, 6.10), # post means
    test_statistics = c("p < 0.01", "p < 0.001"), # test statistics
    effect_size = c(0.517, 1.002) # Effect size
  )
```

```{r echo=FALSE}
pp_2 %>%
  gt() %>%
  tab_header(
    title = "Summary of Test Statistics: Pre-and-post Survey"
  ) %>%
  cols_align(align = "center") %>%
  tab_source_note(
    source_note = "*t-test for PANAS, and Wilcoxon test for ROS"
  )
```

```{r echo=FALSE}
b1 <- data_pre_post_scales %>%
  mutate(clusters=factor(pre_post, levels = c("Pre", "Post"))) %>% 
  select(ROS, pre_post) %>%
  ggplot(aes(x = reorder(pre_post, ROS), y = ROS, fill = pre_post)) +
  geom_boxplot() +
  theme(legend.position = "None") +
  labs(x = "Pre VS Post", y = "ROS")
b2 <- data_pre_post_scales %>%
  mutate(clusters=factor(pre_post, levels = c("Pre", "Post"))) %>%
  select(PANAS, pre_post) %>%
  ggplot(aes(x = reorder(pre_post, PANAS), y = PANAS, fill = pre_post)) +
  geom_boxplot() +
  theme(legend.position = "None") +
  labs(x = "Pre VS Post", y = "PANAS")

b1 + b2
```
